Data Proxy/data-microservice/requirements.txt
aiocoap==0.4.3
asyncio==3.4.3
autopep8==1.6.0
certifi==2022.5.18.1
charset-normalizer==2.0.12
idna==3.3
influxdb-client==1.29.1
msgpack==1.0.3
paho-mqtt==1.6.1
pycodestyle==2.8.0
python-dateutil==2.8.2
pytz==2022.1
requests==2.27.1
Rx==3.2.0
six==1.16.0
toml==0.10.2
urllib3==1.26.9
